Split matrix.cpp into smaller pieces
